Fix backspace key doing nothing at the start of a line
Fix the behavior of the backspace key, wherein when the caret is at
the beginning of an existing line, nothing is done. When pressed with
the caret at the beginning of a line, the backspace key should now
combine the current line and previous line at the previous row.

Close issue #49.