Add a visible caret in the editor
Separate the row data into two elements, 'row-before-caret' and
'row-after-caret'. Modify the handleClick function to check whether
the registered selection is in the 'row-after-caret' element. Add the
length of the 'row-before-caret' element (through the previous caret
offset) to the new offset, if so.

With CSS, add empty content to the 'row-before-caret' element using
the ::after selector and style it as a caret with a blinking
animation.